TowerOfEnzyme

API

TowerOfEnzyme.nth_derivative — Function

nth_derivative(f, x, Val(Order))
nth_derivative(f, x, dx, Val(Order))

Calculates the Order-th directional derivative of f with regards to x in direction dx. If x is a scalar dx can be omitted and is set to one(x).

TowerOfEnzyme.derivative_bundle — Function

derivative_bundle(f, (x, dx, ddx...))

Calculates the higher-order pushforward.

Helper functions

TowerOfEnzyme.jvp — Function

jvp(f, x, v)

Calculate the Jacobian-Vector-Product $J_f(x) * v$

TowerOfEnzyme.hvvp — Function

hvvp(f, x, v)

Calculate the Hessian-Vector-Vector product $H_f(x) * v * v$

Bibliography